refactor: Header 컴포넌트 책임 분리 및 적용
동적으로 Header가 변하므로 이 UI에 대한 책임 분리를 고민해볼 필요가 있다.
