+++ src/layouts/MobileView.tsx
@@ -5,14 +5,18 @@ import Navigation from '../components/@common/Navigation/Navigation';
 import useRouteMatched from '../hooks/useRouteMatch';
 
 const MobileView = () => {
-  const isMatch = useRouteMatched('/profile/:id', '/login', '/write', '/agreement', '/auth', '/feed/:id');
+  const { hasMatched } = useRouteMatched();
+  const isMatchedNavigation = hasMatched('/profile/:id', '/login', '/write', '/agreement', '/auth', '/feed/:id');
+
+  // TODO: Header 도메인 얽힘 문제 해결 확인 시 사용 예정
+  // const isMatchedHeader = hasMatched('/feed', '/feed/:id', '/profile', '/profile/:id', '/profile/:id/more');
 
   return (
     <Container>
       <Wrapper>
         <Outlet />
       </Wrapper>
-      {!isMatch && <Navigation />}
+      {!isMatchedNavigation && <Navigation />}
     </Container>
   );
 };
@@ -30,11 +34,7 @@ const Container = styled.main`
 `;
 
 const Wrapper = styled.div`
-  /* height: calc(100% - 50px); */
   height: 100dvh;
-  /* height: calc(var(--vh, 1vh) * 100); */
-  /* height: -webkit-fill-available; */
-  /* height: fill-available; */
   overflow: auto;
   -ms-overflow-style: none; /* IE and Edge */
   scrollbar-width: none; /* Firefox */
