+++ src/pages/feed/Feeds.page.tsx
@@ -8,6 +8,7 @@ import useCheckbox from '../../components/@common/@hooks/useCheckbox';
 import useRadio from '../../components/@common/@hooks/useRadio';
 import Button from '../../components/@common/Button/Button';
 import CheckboxContainer from '../../components/@common/CheckboxContainer/CheckboxContainer';
+import Header from '../../components/@common/Header/Header';
 import RadioContainer from '../../components/@common/RadioContainer/RadioContainer';
 import Spacer from '../../components/@common/Spacer/Spacer';
 import Text from '../../components/@common/Text/Text';
@@ -15,7 +16,6 @@ import BottomSheet from '../../components/BottomSheet/BottomSheet';
 import FilterBox from '../../components/BottomSheet/FilterBox';
 import IdeaCard from '../../components/Card/IdeaCard';
 import PopCard from '../../components/Card/PopCard';
-import { Header } from '../../components/Header/Header';
 import Dropdown from '../../components/Inputs/Dropdown/Dropdown';
 import Padding from '../../components/Padding';
 import UnStyleButton from '../../components/UnStyleButton';
